Move impl classes; add config/config_file_name to bootstrap
This commit moves all of the implementation-detail classes into
a namespace called `impl` to make it more obvious that there are
no API guarantees for them.

It also adds in the necessary logic to include `:config` and
`:config_file_name` in the "bootstrap" settings, so that if
a user overrides one of those or `:confdir`, the interpolation
will work properly to provide the final value for `:config`.